src/kgmk/sushida/requirements.txt
beautifulsoup4==4.9.3
numpy==1.21.0
Pillow==8.3.1
pyautogui==0.9.52
pytesseract==0.3.8
python-xlib==0.30
PyYAML==5.4.1
requests==2.25.1
selenium==3.141.0
webdriver-manager==3.4.2